+++ libs/id3/src/util/RawId3Frame.ts
@@ -3,4 +3,4 @@
  *
  * @internal
  */
-export type RawId3Frame = { type: string; size: number; data: Uint8Array<ArrayBuffer>; };
+export type RawId3Frame = { type: string; size: number; data: Uint8Array; };