+++ libs/id3/src/util/decodeId3PrivFrame.ts
@@ -13,7 +13,7 @@ import { utf8ArrayToStr } from './utf8ArrayToStr.ts'
  */
 export function decodeId3PrivFrame(
 	frame: RawId3Frame,
-): DecodedId3Frame<ArrayBuffer> | undefined {
+): DecodedId3Frame<ArrayBufferLike> | undefined {
 	/*
 	Format: <text string>\0<binary data>
 	*/